Refactor tests into separate files
The single test file was getting very long, and longer with each new test.
